requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.2

# Computer-vision & model utilities
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
PyYAML==6.0.2
transformers>=4.44
safetensors==0.4.3

# PyTorch stack (install per platform)
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121   # CUDA 12.1
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu      # CPU only
ultralytics==8.3.43
